fix: 修复编译错误和测试 bug
- 移除 types.go 中重复定义的 MailRequestBody、MailRecipient、MailAttachment
- 补充 ExternalRequestBody 类型和 validateExternalRequestBody 函数
- 修复 utils_test.go 中 TestGetErrorCode 的 errcode 类型（int → float64）
- 删除空的遗留文件 wecomchan.go
- 修正 go.mod module 名称

+++ utils_test.go
@@ -81,9 +81,9 @@ func TestGetErrorCode(t *testing.T) {
 		{nil, 0},
 		{map[string]interface{}{}, 0},
 		{map[string]interface{}{"errcode": nil}, 0},
-		{map[string]interface{}{"errcode": 0}, 0},
-		{map[string]interface{}{"errcode": 42001}, 42001},
-		{map[string]interface{}{"errcode": 40001}, 40001},
+		{map[string]interface{}{"errcode": float64(0)}, float64(0)},
+		{map[string]interface{}{"errcode": float64(42001)}, float64(42001)},
+		{map[string]interface{}{"errcode": float64(40001)}, float64(40001)},
 	}
 
 	for _, tt := range tests {
